invoice voucher payment change to type guard
